lang/pydantic-ai-runtime-python/pyproject.toml
[project]
name = "pydantic-ai-runtime"
version = "0.1.0"
description = "EAASP L1 Runtime — Python pydantic-ai agent over OpenAI-compatible endpoints"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.70.0",
    "grpcio-tools>=1.70.0",
    "protobuf>=5.0",
    "pydantic-ai>=0.0.14",
    "pydantic>=2.5",
    "loguru",
    "anyio>=4.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
pydantic-ai-runtime = "pydantic_ai_runtime.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pydantic_ai_runtime"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
